Split AdminController into area-specific classes
Split the admin UI's controller into a bunch of smaller classes, each focused
on a specific subject area.

This is a verbatim cut-and-paste of the existing code, with two exceptions:

- The helper methods for redirecting after POST requests are renamed to clarify
  that they return redirect directives.
- Controller classes whose endpoints are all super-admin-only don't include
  `AcceleratorAdmin` in their lists of required global roles.
